renamed MainMontentComponent file to HomeContent and changed the folder structure to put it under the homeContent folder and reflected that in the imports
