+++ frontend/src/components/mainContent/homeContent/HomeContent.style.tsx
