+++ frontend/src/components/services/Services.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { MainContent } from '../mainContent/MainContentComponent.style';
+import { MainContent } from '../mainContent/homeContent/HomeContent.style';
 
 const Services: React.FC = () => {
     return <MainContent>Services</MainContent>;
